Changes requested by tannewt
