+++ examples/st7735_128x160_simpletest.py
@@ -5,7 +5,7 @@ and draw a solid red background
 
 import board
 import displayio
-from adafruit_st7735 import ST7735R_BLACK
+from adafruit_st7735.st7735r import ST7735R
 
 spi = board.SPI()
 tft_cs = board.D5
@@ -14,7 +14,7 @@ tft_dc = board.D6
 displayio.release_displays()
 display_bus = displayio.FourWire(spi, command=tft_dc, chip_select=tft_cs, reset=board.D9)
 
-display = ST7735R_BLACK(display_bus, width=128, height=160)
+display = ST7735R(display_bus, width=128, height=160, init=b"\x36\x01\xC0")
 
 # Make the display context
 splash = displayio.Group(max_size=10)
